Se culminó con la implementación del US045 y su prueba unitaria. Además, se cambió el tipo de variable para Posición (TablaPosicionesBE) de int a int64 cambió que afectó al frmModuloEntrenamiento en el que se utiliza la variable Posición por lo que se hizo un cast a int en dicho formulario.

+++ trunk/Construcción/UPC.Proyecto.SISPPAFUT/UPC.Proyecto.SISPPAFUT.BL.BC/TablaPosicionesBC.cs
@@ -54,9 +54,9 @@ namespace UPC.Proyecto.SISPPAFUT.BL.BC
             }
         }
 
-        public int ConsultarPosicionEquipoTabla(int codLiga, int codEquipo)
+        public Int64 ConsultarPosicionEquipoTabla(int codLiga, int codEquipo)
         {
-            int pos = 0;
+            Int64 pos = 0;
 
             List<TablaPosicionesBE> lstTabla;
